requirements.txt
streamlit
requests
beautifulsoup4
transformers
python-dotenv
torch
